% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkRetrieve.R
\name{bulkRetrieve}
\alias{bulkRetrieve}
\title{Bulk retrieve information for a list of proteins}
\usage{
bulkRetrieve(protein_list)
}
\arguments{
\item{protein_list}{list of protein identifiers}
}
\value{
a list of S3 objects containing information on the given proteins
}
\description{
The function to bulk retrieve information for a list of proteins.
}
\examples{
orthologs = getData(type="protein",id='YEAST58')$orthologs
bulkRetrieve(orthologs$omaid)
}
