% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{$.omadb_obj}
\alias{$.omadb_obj}
\title{Resolve URLs automatically when accessed}
\usage{
\method{$}{omadb_obj}(x, name)
}
\arguments{
\item{x}{object}

\item{name}{attribute}
}
\value{
API response behind the URL
}
\description{
The function to obtain further information from a given url.
}
