% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHOG.R
\name{childrenMembers}
\alias{childrenMembers}
\title{Get Members of Children HOGs Function}
\usage{
childrenMembers(hog)
}
\arguments{
\item{hog}{a roma hog object}
}
\value{
an list of data frames containing information on members for each child hog
}
\description{
The function to obtain the information available for a Hierarchical orthologous group entry in the datase.
}
\examples{
childrenMembers(getHOG(id = "YEAST590"))
}
