% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{Get the Formatted Data Function}
\usage{
formatData(data)
}
\arguments{
\item{data}{the list to be formatted}
}
\value{
a dataframe
}
\description{
The function to convert a named list into a corresponding dataframe.
}
\examples{
formatData(getData(type="Genome",id="YEAST")$chromosomes)
}
