% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopGO.R
\name{formatTopGO}
\alias{formatTopGO}
\title{Format the GO annotations data}
\usage{
formatTopGO(geneList, format)
}
\arguments{
\item{geneList}{the list of roma protein objects to be included in the analysis - this is where the GO annotations are extracted from}

\item{format}{format for the data to be returned in - either 'GO2geneID' or 'geneID2GO'}
}
\value{
either GO2geneID list or geneID2GO depending on the format parameter
}
\description{
The function to create a list of GO annotations that is compatible with topGO from protein objects in roma
}
\examples{
geneList = list(getData(type="protein",id="YEAST58"),getData(type="protein",id="YEAST00059"))
annotations = formatTopGO(geneList,format="geneID2GO")
}
