% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBioStrings.R
\name{getAAString}
\alias{getAAString}
\title{Get the BioString AAString Object}
\usage{
getAAString(AAString)
}
\arguments{
\item{AAString}{the amino acid string of interest}
}
\value{
an AAString object
}
\description{
The function to obtain the information available for a single entry in the datase.
}
\examples{
getAAString(AAString="NPROK")
}
