% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{getAnnotation-deprecated}
\alias{getAnnotation-deprecated}
\title{Get GO annotation for a sequence Function}
\usage{
getAnnotation(query)
}
\arguments{
\item{query}{the sequence to be annotated, it can be either a string or an AAString object from the Biostrings package}
}
\value{
a data.frame containg the GO annotaition information linked to the query sequence
}
\description{
This function should no longer be used. Use instead \code{\link{annotateSequence}}.
}
\seealso{
\code{\link{OmaDB-deprecated}}
}
\keyword{internal}
