% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romaObject.R
\name{getAttribute}
\alias{getAttribute}
\title{Get the value for the Object Attribute}
\usage{
getAttribute(obj, attribute)
}
\arguments{
\item{obj}{the object of interest}

\item{attribute}{the attribute of interest}
}
\value{
an value for a given object attribute
}
\description{
The function to obtain the value for an object attribute.
}
\examples{
members = getAttribute(getData("group","YEAST58"),'members')
}
