% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get the Data Function}
\usage{
getData(type, id = NULL)
}
\arguments{
\item{type}{the type for the entry to be returned - either protein, genome, group or hog}

\item{id}{an identifier for the entry to be returned. For more information, see the vignette.}
}
\value{
an object containing the JSON keys as attributes
}
\description{
The function to obtain the information available for a single entry in the datase.
}
\examples{
getData(type = "Protein", id="YEAST00001")
getData(type = "Group", id="YEAST00001")
}
