% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBioStrings.R
\name{getDNAString}
\alias{getDNAString}
\title{Get the BioString DNAString Object}
\usage{
getDNAString(DNAString)
}
\arguments{
\item{DNAString}{the DNA string of interest}
}
\value{
an DNAString object
}
\description{
The function to obtain the information available for a single entry in the datase.
}
\examples{
getDNAString(DNAString="ATGC")
}
