% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData2.R
\name{getGenome}
\alias{getGenome}
\title{Get the Genome data function}
\usage{
getGenome(id, attribute = NULL)
}
\arguments{
\item{id}{identifier(s) for the entry or entries to be returned. a character string if single entry or a vector if multiple.}

\item{attribute}{an extra attribute to be returned (proteins)}
}
\value{
an object containing the JSON keys as attributes or a dataframe
}
\description{
The function to obtain the information available for a single protein or multiple proteins in the datase.
}
\examples{
getGenome(id="HUMAN")
getGenome(id="HUMAN",attribute='proteins')
}
