% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{getGenomeAlignment-deprecated}
\alias{getGenomeAlignment-deprecated}
\title{Get Whole Genome Alignment Function}
\usage{
getGenomeAlignment(genome1, genome2, chr1 = NULL, chr2 = NULL,
  rel_type = NULL)
}
\arguments{
\item{genome_id1}{an identifier for the first genome, which can be either its taxon id or UniProt species code}

\item{genome_id2}{an an identifier for the second genome, which can be either its taxon id or UniProt species code}

\item{chr1}{the chromosome of interest for the first genome}

\item{chr2}{the chromosome of interest for the second genome}

\item{rel_type}{the pairs relationship type}

\item{per_page}{the number of instances to be returned or 'all'. default is set to a 100.}
}
\value{
a dataframe containing information about both the entries in the orthologous pair and their relationship
}
\description{
This function should no longer be used. Use instaed \code{\link{getGenomePairs}}.
}
\seealso{
\code{\link{OmaDB-deprecated}}
}
\keyword{internal}
