% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membersDf.R
\name{getGRanges}
\alias{getGRanges}
\title{Get the GRanges object}
\usage{
getGRanges(df)
}
\arguments{
\item{df}{the dataframe or a list of dataframes containing the genomic range data of interest}
}
\value{
an GRanges object
}
\description{
The function to obtain the GRanges object from a dataframe containing a list of members.
}
\examples{
gRanges = getGRanges(df = getData("group","YEAST58")$members)
}
