% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocus.R
\name{getLocus}
\alias{getLocus}
\title{Get loci for a given list of proteins}
\usage{
getLocus(proteins)
}
\arguments{
\item{proteins}{the dataframe or a list of dataframes containing the protein data of interest. this can either be the members df or a list of protein ids.}
}
\value{
genomic range object from the GenomicRanges package in Bioconductor
}
\description{
Function to obtain loci in genomic range format for a given list of proteins
}
\examples{
loci = getLocus(proteins = getOMAGroup("YEAST58")['members'])
}
