% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romaObject.R
\name{getObjectAttributes}
\alias{getObjectAttributes}
\title{Get the Object Attributes}
\usage{
getObjectAttributes(obj)
}
\arguments{
\item{obj}{the object of interest}
}
\value{
an list of object attributes and their data classes
}
\description{
The function to obtain the attributes and their data types for the object created.
}
\examples{
attributes = getObjectAttributes(getData("group","YEAST58"))
}
