% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData2.R
\name{getOMAGroup}
\alias{getOMAGroup}
\title{Get the OMA group data function}
\usage{
getOMAGroup(id, attribute = NULL)
}
\arguments{
\item{id}{identifier(s) for the entry or entries to be returned. a character string if single entry or a vector if multiple.}

\item{attribute}{an extra attribute to be returned (close_groups)}
}
\value{
an object containing the JSON keys as attributes or a dataframe
}
\description{
The function to obtain the information available for a single protein or multiple proteins in the datase.
}
\examples{
getOMAGroup(id="58")
getOMAGroup(id="58",attribute='close_groups')
}
