% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membersDf.R
\name{getOntologies}
\alias{getOntologies}
\title{Get the gene ontologies from members}
\usage{
getOntologies(df)
}
\arguments{
\item{df}{the dataframe or a list of dataframes containing the member proteins of interest}
}
\value{
a list containing the geneID2GO information
}
\description{
The function to obtain the gene ontologies from a dataframe of members.
}
\examples{
ontologies = getOntologies(df = getData("group","YEAST58")$members)
}
