% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membersDf.R
\name{getSequences}
\alias{getSequences}
\title{Get the sequences from members}
\usage{
getSequences(df)
}
\arguments{
\item{df}{the dataframe containing the member proteins of interest}
}
\value{
a list containing the AAString object for each member gene
}
\description{
The function to obtain the protein seqeunces from a dataframe of members.
}
\examples{
sequences = getSequences(df = getData("group","YEAST58")$members)
}
