% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVersion.R
\name{getVersion}
\alias{getVersion}
\title{Get the API and database version function}
\usage{
getVersion()
}
\value{
S3 object
}
\description{
The function to obtain the API and database version that the package is using.
}
\examples{
getVersion()
}
