% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{getXref-deprecated}
\alias{getXref-deprecated}
\title{Get the CrossReferences in the OMA database for a pattern}
\usage{
getXref(pattern)
}
\arguments{
\item{pattern}{the pattern to query the OMA database with - needs to be at least 3 characters long}
}
\value{
a data.frame containing information on the cross references for a given pattern
}
\description{
This function is should no longer be used. Use instead \code{\link{searchProtein}}.
}
\seealso{
\code{\link{OmaDB-deprecated}}
}
\keyword{internal}
