\name{OmaDB-deprecated}
\alias{OmaDB-deprecated}
\title{Deprecated functions in package \sQuote{OmaDB}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{OmaDB} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{getAnnotation: \code{\link{annotateSequence}}}

  }
}
