% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{OmaDB-deprecated}
\alias{OmaDB-deprecated}
\alias{getXref}
\alias{getAnnotation}
\alias{getGenomeAlignment}
\alias{getData}
\title{Deprecated functions in package \pkg{OmaDB}.}
\usage{
getXref(pattern)

getAnnotation(query)

getGenomeAlignment(genome1, genome2, chr1 = NULL, chr2 = NULL,
  rel_type = NULL)

getData(type, id, attribute = NULL)
}
\description{
These functions are provided for compatibility with
older versions of \pkg{OmaDB} only, and will be defunct at the
next release.
}
\details{
The following functions are deprecated and will be
made defunct; use the replacement indicated below:
}
\section{\code{getXref}}{

For \code{getXref}, use \code{\link{searchProtein}}.
}

\section{\code{getAnnotation}}{

For \code{getAnnotation}, use \code{\link{annotateSequence}}.
}

\section{\code{getGenomeAlignment}}{

For \code{getGenomeAlignment}, use \code{\link{getGenomePairs}}.
}

\section{\code{getData}}{

For \code{getData}, use \code{\link{getProtein}, \link{getGenome}, \link{getOMAGroup}}.
}

\keyword{internal}
