% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveURL.R
\name{resolveURL}
\alias{resolveURL}
\title{Get the Further Information behind the URL Function}
\usage{
resolveURL(url_field, pagination = TRUE)
}
\arguments{
\item{url_field}{the url of interest}

\item{pagination}{boolean for which the default is TRUE}
}
\value{
an object containing JSON keys as attributes
}
\description{
The function to obtain further information from a given url.
}
\examples{
resolveURL(url_field="http://omadev.cs.ucl.ac.uk/api/protein/YEAST58/ontology/")
}
