% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roma.R
\docType{package}
\name{roma-package}
\alias{roma}
\alias{roma-package}
\title{roma: A package for the orthology prediction data download from OMA database.}
\description{
roma is a wrapper for the REST API for the Orthologous MAtrix project (OMA) which is a  database for the inference of orthologs among complete genomes. For more details on the OMA project, see \url{https://omabrowser.org/oma/home/}.
}
