% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchProtein.R
\name{searchProtein}
\alias{searchProtein}
\title{Get the CrossReferences in the OMA database for a pattern}
\usage{
searchProtein(pattern)
}
\arguments{
\item{pattern}{the pattern to query the OMA database with - needs to be at least 3 characters long}
}
\value{
a data.frame containing information on the cross references for a given pattern
}
\description{
The function to list all the crossreferences that match a certain defined pattern.
}
\examples{
searchProtein(pattern='MAL')
}
