% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_annotation.R
\docType{data}
\name{sequence_annotation}
\alias{sequence_annotation}
\title{An example dataframe containing GO annotations identified from a given sequence.}
\format{A dataframe with 13 variables:
\describe{
  \item{Qualifier}{qualifier of the annotation}
  \item{GO_ID}{GO term for the annotation}
  \item{With}{GO term for the annotation}
  \item{Evidence}{evidence for the annotation}
  \item{Date}{date}
  \item{DB_Object_Type}{identified object type}
  \item{DB_Object_Name}{identified object name}
  \item{Aspect}{aspect}
  \item{Assigned_By}{assignment of the annotation}
  \item{GO_name}{GO term name}
  \item{DB}{database}
  \item{DB.Reference}{database reference}
  \item{Synonym}{synonym}
  ...
}}
\source{
\url{https://omabrowser.org/api/function/?query=MNDPSLLGYPNVGPQQQQQQQQQQHAGLLGKGTPNALQQQLHMNQLTGIPPPGLMNNSDVHTSSNNNSRQLLDQLANGNANMLNMNMDNNNNNNNNNNNNNNNGGGSGVMMNASTAAVNSIGMVPTVGTPVNINVNASNPLLHPHLDDPSLLNNPIWKLQLHLAAVSAQSLGQPNIYARQNAMKKYLATQQAQQAQQQAQQQAQQQVPGPFGPGPQAAPPALQPTDFQQSHIAEASKSLVDCTKQALMEMADTLTDSKTAKKQQPTGDSTPSGTATNSAVSTPLTPKIELFANGKDEANQALLQHKKLSQYSIDEDDDIENRMVMPKDSKYDDQLWHALDLSNLQIFNISANIFKYDFLTRLYLNGNSLTELPAEIKNLSNLRVLDLSHNRLTSLPAELGSCFQLKYFYFFDNMVTTLPWEFGNLCNLQFLGVEGNPLEKQFLKILTEKSVTGLIFYLRDNRPEIPLPHER}
}
\usage{
sequence_annotation
}
\description{
An example dataframe containing GO annotations identified from a given sequence.
}
\keyword{datasets}
