% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_map.R
\docType{data}
\name{sequence_map}
\alias{sequence_map}
\title{An example dataframe containing proteins identified from a given sequence.}
\format{A dataframe with 3 variables:
\describe{
  \item{query}{sequence that was queried}
  \item{identified_by}{type of identification}
  \item{targets}{list of protein targets identified}
  ...
}}
\source{
\url{https://omabrowser.org/api/sequences/?query=MNDPSLLGYPNVGPQQQQQQQQQQHAGLLGKGTPNALQQQLHMNQLTGIPPPGLMNNSDVHTSSNNNSRQLLDQLANGNANMLNMNMDNNNNNNNNNNNNNNNGGGSGVMMNASTAAVNSIGMVPTVGTPVNINVNASNPLLHPHLDDPSLLNNPIWKLQLHLAAVSAQSLGQPNIYARQNAMKKYLATQQAQQAQQQAQQQAQQQVPGPFGPGPQAAPPALQPTDFQQSHIAEASKSLVDCTKQALMEMADTLTDSKTAKKQQPTGDSTPSGTATNSAVSTPLTPKIELFANGKDEANQALLQHKKLSQYSIDEDDDIENRMVMPKDSKYDDQLWHALDLSNLQIFNISANIFKYDFLTRLYLNGNSLTELPAEIKNLSNLRVLDLSHNRLTSLPAELGSCFQLKYFYFFDNMVTTLPWEFGNLCNLQFLGVEGNPLEKQFLKILTEKSVTGLIFYLRDNRPEIPLPHER}
}
\usage{
sequence_map
}
\description{
An example dataframe containing proteins identified from a given sequence.
}
\keyword{datasets}
