% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setAPI}
\alias{setAPI}
\title{Set the url to the OMA Browser API}
\usage{
setAPI(url)
}
\arguments{
\item{url}{Base url to the API}
}
\description{
Function to set the base url to the OMA Browser API. If no url is
specified, the default OMA Browser API url is used.
}
