% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\docType{data}
\name{taxonomy}
\alias{taxonomy}
\title{An example newick format taxonomy object.}
\format{An S3 with 2 variables:
\describe{
  \item{root_taxon}{sequence that was queried}
  \item{newick}{taxonomy newick}
  ...
}}
\source{
\url{https://omabrowser.org/api/taxonomy/Alveolata/?type=newick}
}
\usage{
taxonomy
}
\description{
An example newick format taxonomy object.
}
\keyword{datasets}
