% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenomePairs.R
\name{getGenomePairs}
\alias{getGenomePairs}
\title{Get Genome Pairs Function}
\usage{
getGenomePairs(genome_id1, genome_id2, chr1 = NULL, chr2 = NULL,
  rel_type = NULL, ...)
}
\arguments{
\item{genome_id1}{an identifier for the first genome, which can be either its taxon id or UniProt species code}

\item{genome_id2}{an an identifier for the second genome, which can be either its taxon id or UniProt species code}

\item{chr1}{the chromosome of interest for the first genome}

\item{chr2}{the chromosome of interest for the second genome}

\item{rel_type}{the pairs relationship type}

\item{...}{qwargs}
}
\value{
a dataframe containing information about both the entries in the orthologous pair and their relationship
}
\description{
The function to obtain the list of orthologs for 2 whole genomes.
}
\examples{
getGenomePairs(genome_id1="YEAST",genome_id2="ASHGO")
}
