% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHOG.R
\name{getHOG}
\alias{getHOG}
\title{Get the HOG Data Function}
\usage{
getHOG(id, level = NULL, members = FALSE)
}
\arguments{
\item{id}{an identifier for the entry to be returned - either its id or one of its protein members}

\item{level}{a specific level for the HOG to be restricted to - set to the root level by default. A taxonomic level can be identified by its full capitalised name e.g. "Fungi" or "Saccharomycetaceae".}

\item{members}{boolean that when set to TRUE returns a dataframe containg the protein members at a given hog and/or level}
}
\value{
an object containing the JSON keys as attributes
}
\description{
The function to obtain the information available for a Hierarchical orthologous group entry in the datase.
}
\examples{
getHOG(id = "YEAST590")
getHOG(id = "YEAST590",level="Saccharomycetaceae", members=TRUE)
}
