% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfo.R
\name{getInfo}
\alias{getInfo}
\title{Get further information for a dataframe of members}
\usage{
getInfo(df, type, format = NULL)
}
\arguments{
\item{df}{the dataframe or a list of dataframes containing the genomic range data of interest}

\item{type}{the type of information to be retrieved}

\item{format}{currently only relevant to type = ontologies where it can be set to either "geneID2GO" or "GO2geneID". Default is "geneID2GO"}
}
\value{
an list
}
\description{
The function to obtain further information from a dataframe containing a list of members.
}
\examples{
sequences = getInfo(df = getData("group","YEAST58")['members'],type='sequences')
}
