% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData2.R
\name{getProtein}
\alias{getProtein}
\title{Get the Protein data function}
\usage{
getProtein(id, attribute = NULL, matchPartially = FALSE)
}
\arguments{
\item{id}{identifier(s) for the entry or entries to be returned. a character string if single entry or a vector if multiple.}

\item{attribute}{an extra attribute to be returned.}

\item{matchPartially}{boolean. If set to TRUE, in case there is no exact match is found in the database it will return top hits.}
}
\value{
an object containing the JSON keys as attributes or a dataframe
}
\description{
The function to obtain the information available for a single protein or multiple proteins in the datase.
}
\examples{
getProtein(id="MAL")
getProtein(id="YEAST00001")
getProtein(id="YEAST00001",attribute='ontology')
getProtein(id=c("YEAST00001","YEAST00002","YEAST00002"))
getProtein(id=c("YEAST00001","YEAST00002","YEAST00002"),attribute='ontology')
}
