% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopGO.R
\name{getTopGO}
\alias{getTopGO}
\title{Get the topGO Object function}
\usage{
getTopGO(annotations, format, myInterestingGenes)
}
\arguments{
\item{annotations}{list of GO annoatations obtained from the formatTopGO()}

\item{format}{format for the data to be returned in - either 'GO2geneID' or 'geneID2GO'}

\item{myInterestingGenes}{list of identifiers for the genes of interest or a dataframe containing them}
}
\value{
topGO object
}
\description{
The function to create a topGO object containing the GO annotations for the given protein list.
}
\examples{
geneList = list(getData(type="protein",id="YEAST58"),getData(type="protein",id="YEAST00059"))
annotations = formatTopGO(geneList,format="geneID2GO")
library(topGO)
getTopGO(annotations, myInterestingGenes = list("YEAST00058"), format = "geneID2GO")
}
