% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTree.R
\name{getTree}
\alias{getTree}
\title{Get the Tree Object}
\usage{
getTree(newick)
}
\arguments{
\item{newick}{The newick tree to be instantiated.}
}
\value{
a tree object
}
\description{
A convenience function to obtain a tree object from newick tree,
essentially wraps read.tree from the ape package.
}
\examples{
taxonomy = getTaxonomy(root="Alveolata")
getTree(newick=taxonomy$newick)
}
