% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXref.R
\name{getXref}
\alias{getXref}
\title{Get the CrossReferences in the OMA database for a pattern}
\usage{
getXref(pattern)
}
\arguments{
\item{pattern}{the pattern to query the OMA database with - needs to be at least 3 characters long}
}
\value{
a data.frame containing information on the cross references for a given pattern
}
\description{
The function to list all the crossreferences that match a certain defined pattern.
}
\examples{
getXref(pattern="MAL")
}
