% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolveURL}
\alias{resolveURL}
\title{Load data for a given url from the OMA Browser API.}
\usage{
resolveURL(url)
}
\arguments{
\item{url}{The url of interest}
}
\value{
a data.frame containing the information behind an URL
}
\description{
This function is usualy not needed by users. In most circumstances
an attribute containing a URL is automatically loaded when accessed.
However, in case the data is transformed into a dataframe, this will
no longer be true, in which case one can access the data behind this
attribute using this function.
}
\examples{
resolveURL("http://omabrowser.org/api/protein/YEAST58/gene_ontology/")
}
