% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveURL.R
\name{resolveURL}
\alias{resolveURL}
\title{Get the Further Information behind the URL Function}
\usage{
resolveURL(url_field)
}
\arguments{
\item{url_field}{the url of interest}
}
\value{
a data.frame containing the information behind an URL
}
\description{
The function to obtain further information from a given url.
}
\examples{
resolveURL(url_field="http://omadev.cs.ucl.ac.uk/api/protein/YEAST58/ontology/")
}
